去除一打开软件就弹出提示框的bug

+++ GradingFund/Stock.cs
@@ -63,7 +63,7 @@
         /// <summary>
         ///     成交额（亿）
         /// </summary>
-        public float Tm;
+        public double Tm;
 
         /// <summary>
         ///     换手率
